+++ CHANGE_LOG.md
@@ -76,6 +76,9 @@ Change log was not kept before version 2.1.
      - Introduced `%thread` format specifier for logging thread ID
      - Explicit check for `__func__` to prevent potential compilation warnings
   * 8.47: Keeps OS evaluated macros defined
+  * 8.48:
+     - Fixes bug with configurating all loggers including TRACE level
+     - Fixes std::thread issue for getting thread ID
 
 #### 7.xx
  * 7.00: